scripts/requirements-gsc.txt
google-api-python-client>=2.170.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
